Update rust version for CI (#238)
* Update rust version for CI

* Fix resource class

* Fix clippy warnings

* Allow lint, should be fixed in the future

* Upgrade chrono

* Upgrade lru

* Upgrade thread_local

* Upgrade zeroize_derive

* Upgrade tracing-subscriber

* Ignore regex advisory

We cannot fix this without upgrading arc

+++ crates/bls-crypto/src/hash_to_curve/try_and_increment_cip22.rs
@@ -90,7 +90,7 @@ where
         let inner_hash = self.hasher.crh(domain, message, hash_bytes)?;
         let mut counter = [0; 1];
         for c in 0..NUM_TRIES {
-            (&mut counter[..]).write_u8(c as u8)?;
+            (&mut counter[..]).write_u8(c)?;
 
             // concatenate the message with the counter
             let msg = &[&counter, extra_data, &inner_hash].concat();
